remote db added

+++ src/main/java/app/repo/CurrencyRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CurrencyRepo extends JpaRepository<Currency, Integer> {
+public interface CurrencyRepo extends JpaRepository<Currency, Long> {
 }
